refactor: move Registry to dokkupy.plugins and mirror test layout
Extract Registry.apply_config into plugins/registry.py and organize
tests under tests/plugins/ to mirror the production package structure.
